pytest/requirements.txt
pytest==7.4.3
pytest-asyncio==0.23.1
pytest-cov==4.1.0
psycopg2-binary==2.9.9
requests==2.31.0
websockets==12.0
pyyaml==6.0.1
